ecore-drm2: Fix output backlight value type definitions
When we go to set output backlight level we use doubles, so lets just
store these values as doubles in the structure.

@fix

+++ src/lib/ecore_drm2/ecore_drm2_private.h
@@ -766,7 +766,7 @@ struct _Ecore_Drm2_Output
    struct
      {
         const char *path;
-        int value, max;
+        double value, max;
         Ecore_Drm2_Backlight_Type type;
      } backlight;
